On an IBM machine with g95 0.93 and gcc 4.3.4 running SLES 11 SP1 I could get the C/Fortran wrappers to work reliably (i.e. they always return gibberish). As a workaround I've rewritten all the wrappers and a few more using the intrinsic C_ISO_BINDING module.
This seems to be working properly, but I could not test it on every possible compiler, furthermore some old/weird
compiler may still not implement the intrinsic module.

So, by default nothing changes, in order to use these new wrappers you have to add __ISO_C_BINDING to
the DFLAGS. I tried to add a check for that in install/configure.am but failed miserably.