dfile_star.f90 will immediately replace open_dvscf_star with a more general interface that allows the dvscf and drho to be opened on the star and to be saved on both the basis of the modes or the cartesian basis. There are probably a few bugs that I'll fix during these week
dfile_autoname.f90 generates automatically a name suitable for storing something dependent on a q vectors from a regular grid in reciprocal space. It also keeps track of the name and allows you to look for the file corresponding to a certain q. A few improvements are necessary for full compatibility with the new wannier-elph code, these will come next.