A few changes to the PAW+EXX subroutines, initially to re-enable the PAW contribution to the Hamiltonian (whic hwas disable by a disguised RETURN statement) and then to fix some bugs it includes.
Notably: the change of the PAW contribution has been changed, this ensures that the energy computed with exxenergy and exxenergy2 are consistent, also when using a sufficiently large cell (or sufficiently good exxdiv_treatment), with spin and a PAW pseudo you can get exactly one Rydberg energy for the Hydrogen atom!

 have also changed electrons.f90 again, and I think it is now satisfactory: during each sub-scf cycle the total energy is printed (but not its components) and you even get the "!" at the end, but at the very final end of the EXX cycle you get a double "!!" at the beginning of the line, which makes grepping less painful.


A few things still missing, to be done sonner or later:
1. US/PAW + noncolinear spin does not work at the moment, because I have not implemented the rotation of the noncolinear <beta|psi> coefficients (this is probably trivial)
2. stress and forces need to be tested
3. stress needs to be tested (it is not implemented with pools either)

I have not really tested band parallelisation.