Fractional translations in "FFT grid" format (ftau) no longer used by pw.x
In my opinion there was a bug in EXX calculations with nonsymmorphic groups: 
ftau was used to rotate \psi_k into \psi_k+q, for usage in Vx\psi, but the 
real-space grid "exx_fft%dfftt" could be different from "dfftp", the grid for
which ftau was computed. Now "ftau" is properly recomputed. Since however 
there is no check on commensurability between "exx_fft%dfftt" and ftau,
the code just stops in that case. It is a quite marginal case, though.